feat(KSegmentedControl): switch to setup syntax [KHCP-6514]
