+++ src/types/segmented-control.ts
@@ -1,0 +1,5 @@
+export interface SegmentedControlOption {
+  label?: string
+  value: string | number | boolean
+  disabled?: boolean
+}
