+++ src/utilities/index.ts
@@ -1,2 +1,3 @@
 export { copyTextToClipboard } from './copyTextToClipboard'
 export { debounce } from './debounce'
+export * from './segmentControlValidators'
