+++ src/utilities/segmentControlValidators.ts
@@ -1,0 +1,38 @@
+import { SegmentedControlOption } from '@/types/segmented-control'
+
+const itemsHaveRequiredProps = (items: SegmentedControlOption[]): boolean => {
+  return items.every(i => i.value !== undefined)
+}
+
+// functions used in prop validators
+const getValues = (items: SegmentedControlOption[]) => {
+  const vals:string[] = []
+  items.forEach((item: SegmentedControlOption) => vals.push(item.value + ''))
+
+  return vals
+}
+
+const itemValuesAreUnique = (items: SegmentedControlOption[]): boolean => {
+  const vals = getValues(items)
+  const uniqueValues = new Set(vals)
+
+  return vals.length === uniqueValues.size
+}
+
+export const validateItems = (items: SegmentedControlOption[] | string[]): boolean => {
+  const isStringArray = typeof items[0] === 'string'
+  const nItems = normalizeItems(items)
+  const isValid = itemValuesAreUnique(nItems)
+
+  return isStringArray ? isValid && itemsHaveRequiredProps(items as SegmentedControlOption[]) : isValid
+}
+
+export const normalizeItems = (items: SegmentedControlOption[] | string[]): SegmentedControlOption[] => {
+  return items.map((item:SegmentedControlOption | string) => {
+    return {
+      label: typeof item === 'string' ? item : (item.label || (item.value + '')),
+      value: typeof item === 'string' ? item : item.value,
+      disabled: typeof item === 'string' ? false : item.disabled,
+    } as SegmentedControlOption
+  })
+}
